+++ src/runtime/components/Modal.vue
@@ -188,7 +188,11 @@ const b24ui = computed(() => tv({ extend: tv(theme), ...(appConfig.b24ui?.modal 
           >
             <div v-if="!!slots.header || (title || !!slots.title) || (description || !!slots.description) || (props.close || !!slots.close)" data-slot="header" :class="b24ui.header({ class: uiProp?.header })">
               <slot name="header" :close="close">
-                <div data-slot="wrapper" :class="b24ui.wrapper({ class: uiProp?.wrapper })">
+                <div
+                  v-if="title || !!slots.title || description || !!slots.description"
+                  data-slot="wrapper"
+                  :class="b24ui.wrapper({ class: uiProp?.wrapper })"
+                >
                   <DialogTitle v-if="!title && !slots.title" />
                   <DialogTitle v-else data-slot="title" :class="b24ui.title({ class: uiProp?.title })">
                     <slot name="title">